% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_average.R
\name{calculate_average}
\alias{calculate_average}
\title{Return an array of average of specific columns matching a pattern}
\usage{
calculate_average(df, pattern)
}
\arguments{
\item{df}{A data frame}

\item{pattern}{A string pattern to match}
}
\value{
A vector of average of specific columns matching a pattern
}
\description{
Return an array of average of specific columns matching a pattern
}
\examples{
df <- data.frame(
    week_payment1 = c(1, 2, 3),
    week_payment2 = c(4, 5, 6),
    week_payment3 = c(7, 8, 9),
    other_column = c(10, 11, 12)
)
calculate_average(df, "week_payment")
}
