% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_creator.R
\name{calculate_percentage_change}
\alias{calculate_percentage_change}
\title{Calculate percentage change over a time period
(months over months or weeks over weeks)
for the given column pattern.}
\usage{
calculate_percentage_change(
  df,
  pattern,
  compare_period = c(2, 2),
  time_filter = None
)
}
\arguments{
\item{df}{A tibble or Data frame}

\item{pattern}{A string representing column pattern}

\item{compare_period}{A numeric vector of comparison period}

\item{time_filter}{A numeric vector of time filter}
}
\value{
A vector with calculated percentage change
}
\description{
Use case:
This function aims to generate features to capture trend
in data for a given comparison period. Example:
Telecommunication - Predict customers who are more likely to decline their revenue next month/week.
Finance - Predict customers who are going to default next month/week.
}
