% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_standard_deviation.R
\name{calculate_standard_deviation}
\alias{calculate_standard_deviation}
\title{Return a vector of standard deviation of specific columns matching a pattern}
\usage{
calculate_standard_deviation(df, pattern)
}
\arguments{
\item{df}{A data frame}

\item{pattern}{A string pattern to match}
}
\value{
A vector of standard deviation of specific columns matching a pattern
}
\description{
Return a vector of standard deviation of specific columns matching a pattern
}
\examples{
df <- data.frame(
week_payment1 = c(0, 3, 6),
week_payment2 = c(2, 5, 8),
week_payment3 = c(5, 8, 11),
other_column = c(10, 11, 12)
)
calculate_standard_deviation(df, "week_payment")
c(6, 6, 6)
}
