% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matching_column_names.R
\name{get_matching_column_names}
\alias{get_matching_column_names}
\title{Returns a subset of the columns whose names match the pattern.}
\usage{
get_matching_column_names(df, pattern)
}
\arguments{
\item{df}{The dataframe from which to extract columns}

\item{pattern}{The prefix of the column names to extract (string)}
}
\value{
A vector of strings that match the pattern
}
\description{
Matching columns are columns whose names start
with the given pattern and end with an incrementing integer.
}
\examples{
data <- data.frame(week_payment1 = c(1, 2, 3), week_payment2 = c(1, 2, 3),
                   week_payment3 = c(1, 2, 3), othercolumn = c(5, 6, 7))
get_matching_column_names(data, "week_payment")
}
