% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_percentage_change.R
\name{calculate_percentage_change}
\alias{calculate_percentage_change}
\title{Calculate percentage change.}
\usage{
calculate_percentage_change(
  df,
  pattern,
  compare_period = c(2, 2),
  time_filter = NULL
)
}
\arguments{
\item{df}{A tibble or Data frame}

\item{pattern}{A string representing column pattern}

\item{compare_period}{A numeric vector of comparison period}

\item{time_filter}{A numeric vector of time filter}
}
\value{
A vector with calculated percentage change
}
\description{
Calculate percentage change over a time period for the given column pattern.
}
\details{
This function aims to generate features such as percentage change in last month versus
last to last month to capture trend in data for a given comparison period.
For example, In Telecommunication it can be used to generate features useful for
predicting subscribers who are more likely to decline their revenue next month/week.
In Finance it can be used to generate features useful for predicting
customers who are going to default next month/week.
}
\examples{
df <- data.frame(
    week_payment1 = c(1, 2, 3),
    week_payment2 = c(4, 5, 6),
    week_payment3 = c(7, 8, 9),
    week_payment4 = c(10, 11, 12)
)
calculate_percentage_change(df, "week_payment")
}
